pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-btc-bot"
version = "1.0.0"
description = "Automated trading bot for Polymarket BTC 5-minute markets"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Abhijeet Maker", email = "abhijeet@example.com"}]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "py-clob-client>=0.1.2",
    "websocket-client>=1.6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/abhijeet-maker/polymarket-btc-bot"
Repository = "https://github.com/abhijeet-maker/polymarket-btc-bot.git"
Issues = "https://github.com/abhijeet-maker/polymarket-btc-bot/issues"